pyproject.toml
[project]
name = "soxray"
version = "0.1.0"
description = ""
authors = [
    {name = "munimdev",email = "munimzafar100@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "reportlab (>=4.2.0,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langgraph (>=1.0.10,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.5"
mypy = "^1.19.1"
pandas-stubs = "^2.2.0"

